multiple login

+++ routes/web.php
@@ -16,14 +16,14 @@ use App\Http\Controllers\ProductController;
 Route::get('/', function(){
     return view('welcome');
 });
-Route::get('/products', [ProductController::class, 'index']);
-Route::get('/products/create', [ProductController::class, 'create']);
-Route::post('/products', [ProductController::class, 'store']);
-Route::get('/products/{id}/edit', [ProductController::class, 'edit']);
-Route::put('/products/{id}', [ProductController::class, 'update']);
-Route::delete('/products/{id}', [ProductController::class, 'destroy']);
-
 Auth::routes();
+Route::get('/products', [ProductController::class, 'index'])->middleware('IsAdmin');
+Route::get('/products/create', [ProductController::class, 'create'])->middleware('IsAdmin');
+Route::post('/products', [ProductController::class, 'store'])->middleware('IsAdmin');
+Route::get('/products/{id}/edit', [ProductController::class, 'edit'])->middleware('IsAdmin');
+Route::put('/products/{id}', [ProductController::class, 'update'])->middleware('IsAdmin');
+Route::delete('/products/{id}', [ProductController::class, 'destroy'])->middleware('IsAdmin');
+
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 Route::get('/adminpage', [App\Http\Controllers\HomeController::class, 'adminPage'])->name('adminpage')->middleware('IsAdmin');
